# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_blueship581_gbinsureapi_internal_dto.SubmitSettlementRequest:
    properties:
      presettlement_id:
        type: integer
      request_no:
        description: 调用方请求流水号（HIS 超时重试保持不变），按调用方做幂等；同号重试返回原
          settlement_no 与状态
        maxLength: 64
        type: string
    required:
    - presettlement_id
    - request_no
    type: object
...
  github_com_blueship581_gbinsureapi_internal_util.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
host: localhost:19935
info:
  contact: {}
  description: '为医院信息系统提供标准化的医保结算接口服务：参保人身份核验、费用上传、预结算、正式结算、结算单查询与日终对账。业务接口需 API Key（X-API-Key）与
    JWT（Authorization: Bearer）双认证。'
  title: 医保智能结算 API 网关
  version: "1.0"
paths:
...
  /api/v1/settlements:
    get:
      parameters:
      - description: 状态
        in: query
        name: status
        type: string
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页数量
        in: query
        name: page_size
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_blueship581_gbinsureapi_internal_util.Response'
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: 结算单列表
      tags:
      - settlements
  /api/v1/settlements/{settlement_no}:
    get:
      parameters:
      - description: 结算单号
        in: path
        name: settlement_no
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_blueship581_gbinsureapi_internal_util.Response'
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: 结算单详情
      tags:
      - settlements
  /api/v1/settlements/{settlement_no}/reverse:
    post:
      description: 结算当日全额回退
      parameters:
      - description: 结算单号
        in: path
        name: settlement_no
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_blueship581_gbinsureapi_internal_util.Response'
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: 结算冲正
      tags:
      - settlements
  /api/v1/settlements/submit:
    post:
      consumes:
      - application/json
      description: 确认预结算后生成唯一结算单号并返回凭证信息
      parameters:
      - description: 预结算 ID
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_blueship581_gbinsureapi_internal_dto.SubmitSettlementRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_blueship581_gbinsureapi_internal_util.Response'
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: 正式结算
      tags:
      - settlements
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: X-API-Key
    type: apiKey
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"